requirements.txt
gunicorn==20.0.4
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Mail
itsdangerous==1.1.0
Jinja2==2.11.2
numpy==1.19.1
opencv-contrib-python-headless
pandas==1.1.0
pytesseract
python-dateutil
requests==2.18.4
Werkzeug==1.0.1
Pillow
fuzzywuzzy
# Flask-PyMongo
pymongo
dnspython
datetime
pytz
python-Levenshtein